requirements-mlx.txt
# Qwen2.5-VL Fine-tuning with MLX-VLM (Native MLX)
# For Apple Silicon (M1/M2/M3/M4) - macOS only
# Install with: pip install -r requirements-mlx.txt
# Or with uv: uv pip install -r requirements-mlx.txt

# Core MLX dependencies
mlx>=0.20.0
mlx-vlm>=0.3.5

# Image processing
pillow>=10.0.0

# Data processing
datasets>=2.14.0
scikit-learn>=1.3.0
tqdm>=4.66.0

# Jupyter notebook support (optional)
# notebook>=7.0.0
# ipywidgets>=8.0.0
